requirements.txt
apprise
loguru
paho-mqtt
pydantic
pydantic-settings